Ignore resource type in BaseResourcePacker.pack_resource
Resource type can be determined at runtime when calling
BaseResourcePacker.pack_resource, so there is no need to
specify it explicitly when instantiating a packer that
will only be used to pack resources. This allows to
correctly pack resources whose type is different from
controller's resource type without explicitly specifying it.
This is useful when returning a list of objects of TypeA
from controller with resource type TypeB.

Change-Id: I1a560470ac38812997fe663df335fec360fb037d